just break out the modules folder into processors and providers
