+++ plugins/catalog-backend/src/index.ts
@@ -20,9 +20,10 @@
  * @packageDocumentation
  */
 
-export * from './modules';
+export * from './processors';
 export * from './processing';
 export * from './search';
 export * from './service';
 export * from './deprecated';
 export * from './constants';
+export * from './util';
