+++ plugins/catalog-backend/src/processing/index.ts
@@ -15,6 +15,5 @@
  */
 
 export type { CatalogProcessingEngine } from './types';
-
 export { createRandomProcessingInterval } from './refresh';
 export type { ProcessingIntervalFunction } from './refresh';
