+++ plugins/catalog-backend/src/processors/AnnotateLocationEntityProcessor.test.ts
