+++ plugins/catalog-backend/src/processors/AnnotateLocationEntityProcessor.ts
