+++ plugins/catalog-backend/src/processors/AnnotateScmSlugEntityProcessor.test.ts
