+++ plugins/catalog-backend/src/processors/AnnotateScmSlugEntityProcessor.ts
