+++ plugins/catalog-backend/src/processors/BuiltinKindsEntityProcessor.test.ts
