+++ plugins/catalog-backend/src/processors/BuiltinKindsEntityProcessor.ts
