+++ plugins/catalog-backend/src/processors/FileReaderProcessor.test.ts
