+++ plugins/catalog-backend/src/processors/FileReaderProcessor.ts
