+++ plugins/catalog-backend/src/processors/LocationEntityProcessor.test.ts
