+++ plugins/catalog-backend/src/processors/LocationEntityProcessor.ts
