+++ plugins/catalog-backend/src/processors/__fixtures__/fileReaderProcessor/awsS3/awsS3-mock-object.txt
