+++ plugins/catalog-backend/src/processors/__fixtures__/fileReaderProcessor/component.yaml
