+++ plugins/catalog-backend/src/processors/__fixtures__/fileReaderProcessor/dir/api.yaml
