+++ plugins/catalog-backend/src/service/CatalogBuilder.ts
@@ -52,17 +52,17 @@ import {
   FileReaderProcessor,
   PlaceholderProcessor,
   UrlReaderProcessor,
-} from '../modules';
-import { ConfigLocationEntityProvider } from '../modules/core/ConfigLocationEntityProvider';
-import { DefaultLocationStore } from '../modules/core/DefaultLocationStore';
+} from '../processors';
+import { ConfigLocationEntityProvider } from '../providers/ConfigLocationEntityProvider';
+import { DefaultLocationStore } from '../providers/DefaultLocationStore';
 import { RepoLocationAnalyzer } from '../ingestion/LocationAnalyzer';
 import { AuthorizedLocationAnalyzer } from './AuthorizedLocationAnalyzer';
 import {
   jsonPlaceholderResolver,
   textPlaceholderResolver,
   yamlPlaceholderResolver,
-} from '../modules/core/PlaceholderProcessor';
-import { defaultEntityDataParser } from '../modules/util/parse';
+} from '../processors/PlaceholderProcessor';
+import { defaultEntityDataParser } from '../util/parse';
 import {
   CatalogProcessingEngine,
   createRandomProcessingInterval,
